Distributed_Event_Management_System/back-end/ticketing/requirements.txt
flask
boto3
Werkzeug
Flask-RESTful
flask-socketio
flask-cors
python-dotenv